+++ mod-pubsub-server/src/main/java/org/folio/services/PublishingService.java
@@ -1,0 +1,19 @@
+package org.folio.services;
+
+import io.vertx.core.Future;
+import org.folio.rest.jaxrs.model.Event;
+
+/**
+ * Publishing Service
+ */
+public interface PublishingService {
+
+  /**
+   * Publishes event to the appropriate topic in kafka
+   *
+   * @param event    event to publish
+   * @param tenantId tenant id
+   * @return future with true if succeeded
+   */
+  Future<Boolean> publishEvent(Event event, String tenantId);
+}